[DataSet,DataArray] Merge classes with mixins
